requirements.txt
# requirements.txt
PyMuPDF
ebooklib
customtkinter
sv-ttk
Pillow
beautifulsoup4
lxml